Minggu ke-7
Video Lumen Video 11-20

+++ api-lumen/database/migrations/2021_02_08_120807_create_menus_table.php
@@ -15,10 +15,10 @@ class CreateMenusTable extends Migration
     {
         Schema::create('menus', function (Blueprint $table) {
             $table->increments('idmenu');
-            $table->integer('idkategori');
+            $table->bigInteger('idkategori');
             $table->string('menu');
             $table->string('gambar');
-            $table->float('harga');
+            $table->bigInteger('harga');
             $table->timestamps();
         });
     }
